router stablished for otp and other authorization and authentication system
