+++ src/utils/asyncHandler.utils.js
@@ -1,0 +1,7 @@
+const asyncHandler = (requestHandler) =>{
+    return (req,res,next)=>{
+        Promise.resolve(requestHandler(req,res,next)).catch((err)=>next(err))
+    }
+}
+
+export default asyncHandler;
